% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{CARM-package}
\alias{CARM-package}
\title{CARM:Covariate-adjusted Adaptive Randomization via Mahalanobis-distance}
\description{
The CARM package provides function of implement of randomization:
}
\section{ARM functions}{

please ?ARM and ?ARMM to view function usage
}

